+++ Engine/General/Graphics/IndexArray.cpp
@@ -1,11 +1,11 @@
 #include "pch.h"
 #include "IndexArray.h"
 
-IndexArray::IndexArray(size_t indexCount, const uint16_t* indexData)
+IndexArray::IndexArray(size_t indexCount, const uint32_t* indexData)
 	:m_indexCount(indexCount)
 {
-	m_indexData = new uint16_t[indexCount];
-	memcpy(m_indexData, indexData, sizeof(uint16_t) * indexCount);
+	m_indexData = new uint32_t[indexCount];
+	memcpy(m_indexData, indexData, sizeof(uint32_t) * indexCount);
 }
 
 IndexArray::~IndexArray()
